Make entry_id equal to Wikipedia edit it.

+++ frontend/schema.sql
@@ -2,7 +2,7 @@ DROP TABLE IF EXISTS plgov_entries;
 
 
 CREATE TABLE plgov_entries (
-  entry_id MEDIUMINT NOT NULL AUTO_INCREMENT,
+  entry_id INT NOT NULL AUTO_INCREMENT,
   title TEXT NOT NULL,
   ip TEXT NOT NULL,
   rdns TEXT,
@@ -13,7 +13,7 @@ CREATE TABLE plgov_entries (
 
 DROP TABLE IF EXISTS plgov_ips;
 CREATE TABLE plgov_ips (
-  entry_id MEDIUMINT NOT NULL,
+  entry_id INT NOT NULL,
   ip VARCHAR(16) NOT NULL,
   CONSTRAINT entry_id_fk
   FOREIGN KEY entry_id_fk(entry_id) REFERENCES plgov_entries (entry_id)
